Update SQL Script 086_to_087

+++ CMI/Access/Sql.Lesesaal/SqlDbScripts/0086_TO_0087.sql
@@ -2,7 +2,7 @@
 /* Add table "SyncAction"                                                 */
 /* ---------------------------------------------------------------------- */
 CREATE TABLE [SyncAction] (
-    [SyncActionId] BIGINT IDENTITY(1,1) NOT NULL,
+    [SyncActionId] INTEGER IDENTITY(1,1) NOT NULL,
     [ArchiveRecordId] NVARCHAR(255),
     [ActionType] NVARCHAR(40),
     [ActionStatus] INTEGER DEFAULT 0,
@@ -43,8 +43,8 @@ GO
 
 
 CREATE TABLE [SyncActionLog] (
-    [SyncActionLogId] BIGINT IDENTITY(1,1) NOT NULL,
-    [SyncActionId] BIGINT,
+    [SyncActionLogId] INTEGER IDENTITY(1,1) NOT NULL,
+    [SyncActionId] INTEGER,
     [LogDate] DATETIME2 DEFAULT getdate(),
     [ActionStatusHistory] NVARCHAR(40),
     [ErrorReason] NVARCHAR(max),
